requirements.txt
pandas
numpy
python-dotenv
MetaTrader5
requests
matplotlib
fastapi
uvicorn
jinja2
websockets
torch
